xcode project updated to xcode5

+++ rendersystems/OpenGL/OpenGL_Texture.cpp
@@ -180,7 +180,7 @@ namespace april
 		if (image != NULL)
 		{
 #if TARGET_OS_IPHONE
-			if (image->format == GL_COMPRESSED_RGBA_PVRTC_4BPPV1_IMG || image->format == GL_COMPRESSED_RGBA_PVRTC_2BPPV1_IMG)
+			if (((unsigned int) image->format) == GL_COMPRESSED_RGBA_PVRTC_4BPPV1_IMG || ((unsigned int) image->format) == GL_COMPRESSED_RGBA_PVRTC_2BPPV1_IMG)
 			{
 				glCompressedTexImage2D(GL_TEXTURE_2D, 0, image->format, image->w, image->h, 0, image->compressedSize, image->data);
 				this->format = FORMAT_ARGB; // TODO - not really a format
